Rename to com.vaadin.flow.component.shared
