+++ vaadin-flow-components-shared-parent/vaadin-flow-components-base/src/main/java/com/vaadin/flow/component/shared/HasThemeVariant.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.flow.component.shared.base;
+package com.vaadin.flow.component.shared;
 
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
