+++ vaadin-text-field-flow-parent/vaadin-text-field-flow/src/main/java/com/vaadin/flow/component/textfield/TextFieldVariant.java
@@ -15,7 +15,7 @@
  */
 package com.vaadin.flow.component.textfield;
 
-import com.vaadin.flow.component.shared.base.ThemeVariant;
+import com.vaadin.flow.component.shared.ThemeVariant;
 
 /**
  * Set of theme variants applicable for the {@link TextField} component, as well
